snake-diffusion/requirements.txt
click
torch
torchvision
matplotlib
einops
PyYAML
pygame
numpy
tqdm
ipywidgets